# Fix aggregate_downsample() performance degradation