--- a/docs/changes/timeseries/13069.bugfix.rst
+++ b/docs/changes/timeseries/13069.bugfix.rst
@@ -1,0 +1,2 @@
+Fixed a performance regression in ``timeseries.aggregate_downsample``
+introduced in Astropy 5.0 / #11266.