@orionlee 2022-04-05
Concern so far. If someone is better equipped to push forward, feel free to takeover. 

1. The fix is done by passing plain array to `np` functions with `.value` instead of `Time` objects, e.g., 

``` python
    indices = np.searchsorted(bin_end.value, ts_sorted.time[keep].value)
```

- 2 `Time.value` instances are not always comparable, if they are in different format, e.g,
  - time1 is of type `isot` , e.g., '2016-03-22T12:30:31', and time2 is `jd` (number)
  - time1 is `jd` , and time2 is `mjd`
  - I'm unsure what'd be a good way to resolve this. Do we convert the `Time` objects to a single format?

---

2. The PR failed 1 assertion in `test_downsample()`. I cannot figure out what the problem is.

https://github.com/astropy/astropy/blob/f54df76430eac0f67efb1e61be5a31f4c3029b9d/astropy/timeseries/tests/test_downsample.py#L127

```python
>       assert_equal(down_1["a"].data.data, np.array([1, 2, 3, 4, 5]))
E       AssertionError:
E       Arrays are not equal
E
E       Mismatched elements: 4 / 5 (80%)
E       Max absolute difference: 5
E       Max relative difference: 1.
E        x: array([1, 3, 4, 5, 0])
E        y: array([1, 2, 3, 4, 5])

astropy\timeseries\tests\test_downsample.py:127: AssertionError
```