@dhomeier 2022-04-05
Thanks for working on this!

> * I'm unsure what'd be a good way to resolve this. Do we convert the `Time` objects to a single format?

Yes, working on a common scale is mandatory, but range and precision has also to be taken into consideration. One might just use `time.jd`, `time.mjd` or `time.unix` to work with, but timeseries may well contain data at millisecond resolution, which would already hit `float64` resolution limits in these cases (`Time.now().jd` ~ 2.1e14 ms; for `unix` it is a bit better, but still not ideal).
I think the safest way will be to use all times relative to the first point in the series, just as done in the previous implementation:
https://github.com/aaryapatil/astropy/blob/cbf4d021b0e4b5ec9aa804cf4ce1145c19e0cf84/astropy/timeseries/downsample.py#L77

This will require some more bookkeeping, but for just the `ts_sorted` and `time_bin_start`/`_end` values hopefully does not become too complicated.