@orionlee 2022-04-06
@dhomeier Would it make sense to encapsulate all such logic (creating relative time and sort them) in a new `Time.searchsorted(v)` function (assuming that making `np.searchsorted()` run fast with `Time` is non-trivial)?