@dhomeier 2022-04-06
> @dhomeier Would it make sense to encapsulate all such logic (creating relative time and sort them) in a new `Time.searchsorted(v)` function (assuming that making `np.searchsorted()` run fast with `Time` is non-trivial)?

Given that there is a speed-up of O(1.e3) involved, I think a general fast search sorted for Time would certainly be worth it.
I don't know exactly what it would involve to make this available through `np.searchsorted`, but it looks like it could build on #12703. The logic for changing to relative and identical timescales would still need to be written, but that should be about the same effort as putting it into a `Time.searchsorted()` – note also @mhvk's comments on precision in that PR.
So I think once @dstansby's PR is merged this could go into a follow-up, implementing `searchsorted` in `CUSTOM_FUNCTIONS` analogously to `linspace`.