@orionlee 2022-04-06
The `delete_indices = np.where(np.logical_and(ts_sorted.time > bin_end[ind], ts_sorted.time < bin_start[ind+1]))` does pose some problem, especially for large data set.
- if I replace `Time` object comparison with plain ndarray by using `.value`, it cuts down some time
- if we remove the check altogether (over `n_bins` times per `downsample` call), the time is further reduced to be quite close to the time in astropy 4.x 

Here is the breakdown using the test script in #13058
|                                                | 20k samples | 200k samples |   |
| ---------------------------------------------- | ----------- | ------------ | -- | 
| astropy 4.3.1                                  | ~0.42sec    | ~4.2sec      |  |
| astropy 5.0.4                                  | ~10sec      | ~252sec      | |
| searchsorted() fix (this PR)                   | ~1.3sec     | 147sec       | <-- **still very slow!!** | 
| `delete_indices` : use ndarray rather than `Time` for comparison | ~1.0sec     | ~14sec       | |
| remove `delete_indices` logic                   | ~0.55sec    | ~5sec        | |

---

I have done [a separate test](https://gist.github.com/orionlee/6b3e384b0b272cd1384a384ead7bde7b) for the running time of `Time` comparison logic. With 200k samples and 40k bins,
- `Time` object comparison: ~71sec
- `ndarray` comparison: ~3.7seec

--

For this PR
- I could replace `Time` comparison with ndarray, though it seems to be something that could better be done with `Time` itself.
- `delete_indices` check logic would remain to be a source of noticeable slow down.