@dhomeier 2022-04-06
Given the large speedup (I only saw something like a factor 2, but only experimented with referencing `Time.unix` directly in the loop), it should be worth defining `bin_start_sec = (bin_start - ts_sorted.time[0])to_value(u.s)` etc. outside of the loop, and then use the optimised loop as `np.where(bin_start_sec[1:] > bin_end_sec[:-1])[0]` (`+ epsilon`).
And one could think about setting something like `1e-3 * (ts_sorted.time[1:]-ts_sorted.time[:-1]).min()` as the epsilon to avoid most of those cases entirely.