@orionlee 2022-04-07
A side note on performance improvement option I just noticed that is out of scope of this PR, but I'm mentioning to see if we should open a separate issue to track / tackle it.

If [`bottleneck`](https://github.com/pydata/bottleneck)'s `nanmedian` is used in the test, the running time is further reduced (200k timeseries: from ~5sec to ~1.6sec ; 20k series from ~0.5sec to 0.2sec).

[`astropy.stats.sigma_clip()`](https://docs.astropy.org/en/stable/api/astropy.stats.sigma_clip.html#astropy.stats.sigma_clip) API is set such that by default if `bottleneck` is installed, it will be used (and typically result in reduced running time) to calculate median/mean, similar approach can be used here.

See: https://docs.astropy.org/en/stable/stats/index.html#performance-tips